Add type hints (#59)
* Introduce Type Hints

* Extend example with custom handlers

* Extend type hints to support options map

The example in `options_map_func.py` shows that the intention of the
custom handler `options_map`func` is to make arbitrary objects
printable. Thus, the type annotations had to be extended to support this
usecase.

* Extend type hint to allow options_map_func to return None

* Run mypy in test stage

* Always return List of selections

* Fix type annotation for options_map_func

* Simplify code by making options_map_func mandatory

Instead of having an optional function the function is mandatory now but
has a default value. This allows to drop a case distinction and, much
more importantly, drop many of the alternatives of the type hints.

* Adapt examples for returned lists

The examples still tried to unpack the returned list. This was not a
good demonstration of the new, streamlined, return style type.